fix: race on create and stats
Volume metadata is created ahead of the img file, which means a concurrent
capacity call may find an error.

Also there's no point supressing file not found only to then hit a key error
exception.
I'm not sure if the metadata management code needs to rely on this somehow,
so keep this behaviour there but change it elsewhere.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>