fix(block): skip node expand if not staged
If a block volume is not staged (if the loop device is not mounted),
then we don't need the node expand to be called before completing
the resize.

Since the controller expand is using a helper pod, and there's no
other means of exchanging information, we use the stdout of the
helper pod like so:
openebs.io/rawfile/is-attached=(True|False)

If the volume is not attached, and it's a block then we're done and
we return node_expansion_required as False.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>